refactor(travis): Move node schedule computations to versioning code
